Changed int type for libboardgame_mcts::PlayerInt and libpentobi_base::Color to uint_fast8_t

+++ src/libpentobi_base/Color.h
@@ -20,7 +20,7 @@ using libboardgame_util::Exception;
 class Color
 {
 public:
-    typedef unsigned IntType;
+    typedef uint_fast8_t IntType;
 
     class InvalidString
         : public Exception
